requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-dotenv
requests
feedparser
beautifulsoup4